backend/requirements.txt
django==3.2.16
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-storages==1.13.1
django-ckeditor==6.3.2
psycopg2==2.9.5
pillow==10.4.0
django-ckeditor==6.3.2